backend/requirement.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session
alembic
flask-cors
flask-jwt-extended
Flask-Mail
itsdangerous
Flask-Migrate